changed then namespace fo the Models
