--- app/UserProject.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App;
-
-use Illuminate\Database\Eloquent\Model;
-
-class UserProject extends Model
-{
-    //
-}
